feat: update speakers (#93)
