+++ constants/index.ts
@@ -1,5 +1,3 @@
-import type { Speaker } from '~/types/speakers'
-
 export const appName = '2024 WebConf Taiwan 技術研討會'
 export const appDescription = '本屆榮幸邀請到各領域專業講者，包含網站開發、前端設計、UIUX 以及團隊管理、軟性主題等各方面相關議題分享。希望能讓所有對 Web 開發、設計有熱情的人，共同創造一個充滿回憶與影響力的盛會！'
 export const ogImageLink = 'https://webconf.tw/og-image.webp'
@@ -40,179 +38,6 @@ export const navLinks = [
     name: '歷屆',
     engName: 'HISTORY',
     href: '#',
-  },
-]
-
-export const speakers: Speaker[] = [
-  {
-    id: 'speaker_1',
-    displayName: 'Mosky',
-    avatar: '/speakers/avatar_mosky.png',
-  },
-  {
-    id: 'speaker_2',
-    displayName: '郭藺瑩 Lydia',
-    avatar: '/speakers/avatar_lydia.png',
-  },
-  {
-    id: 'speaker_3',
-    displayName: '張文鈿 iHower',
-    avatar: '/speakers/avatar_ihower.png',
-  },
-  {
-    id: 'speaker_4',
-    displayName: 'PJwang',
-    avatar: '/speakers/avatar_pjwang.png',
-  },
-  {
-    id: 'speaker_5',
-    displayName: 'Kuro',
-    avatar: '/speakers/avatar_kuro.png',
-  },
-  {
-    id: 'speaker_6',
-    displayName: '李德俊 (紅豆)',
-    avatar: '/speakers/avatar_djlee.png',
-  },
-  {
-    id: 'speaker_7',
-    displayName: '高玉璁 Samuel',
-    avatar: '/speakers/avatar_samuel.png',
-  },
-  {
-    id: 'speaker_8',
-    displayName: '林沅霖 Yuanlin Lin',
-    avatar: '/speakers/avatar_yuanlin.png',
-  },
-  {
-    id: 'speaker_9',
-    displayName: 'Summer',
-    avatar: '/speakers/avatar_summer.png',
-  },
-  {
-    id: 'speaker_10',
-    displayName: '設計師勞哥',
-    avatar: '/speakers/avatar_maylogger.png',
-  },
-  {
-    id: 'speaker_11',
-    displayName: '王偉任 Weithenn',
-    avatar: '/speakers/avatar_weithenn.png',
-  },
-  {
-    id: 'speaker_12',
-    displayName: 'Zet',
-    avatar: '/speakers/avatar_zet.png',
-  },
-  {
-    id: 'speaker_13',
-    displayName: '奶綠茶',
-    avatar: '/speakers/avatar_milkmidi.png',
-  },
-  {
-    id: 'speaker_14',
-    displayName: '李建杭 Amos',
-    avatar: '/speakers/avatar_amos.png',
-  },
-  {
-    id: 'speaker_15',
-    displayName: 'Taiming',
-    avatar: '/speakers/avatar_taiming.png',
-  },
-  {
-    id: 'speaker_16',
-    displayName: '在地上滾的工程師 Nic',
-    avatar: '/speakers/avatar_nic.png',
-  },
-  {
-    id: 'speaker_17',
-    displayName: '黃睿霖 Rei',
-    avatar: '/speakers/avatar_huangruilin.png',
-  },
-  {
-    id: 'speaker_18',
-    displayName: 'Anthony Fu',
-    avatar: '/speakers/avatar_antfu.png',
-  },
-  {
-    id: 'speaker_19',
-    displayName: 'GUGU',
-    avatar: '/speakers/avatar_gugu.png',
-  },
-  {
-    id: 'speaker_20',
-    displayName: 'Leo 李歐',
-    avatar: '/speakers/avatar_leo.png',
-  },
-  {
-    id: 'speaker_21',
-    displayName: '李郁欣',
-    avatar: '/speakers/avatar_sylviauk.png',
-  },
-  {
-    id: 'speaker_22',
-    displayName: '楊時鈞 Steven Yeo',
-    avatar: '/speakers/avatar_stevenyeo.png',
-  },
-  {
-    id: 'speaker_23',
-    displayName: '史蒂夫·葉',
-    avatar: '/speakers/avatar_stevenyeh.png',
-  },
-  {
-    id: 'speaker_24',
-    displayName: '郭丞哲 Doppler kuo',
-    avatar: '/speakers/avatar_doppler.png',
-  },
-  {
-    id: 'speaker_25',
-    displayName: '甘韋',
-    avatar: '/speakers/avatar_weikan.png',
-  },
-  {
-    id: 'speaker_26',
-    displayName: '魏銘信 Ivan Wei',
-    avatar: '/speakers/avatar_ivanwei.png',
-  },
-  {
-    id: 'speaker_27',
-    displayName: '劉宥筠 Iris',
-    avatar: '/speakers/avatar_iris.png',
-  },
-  {
-    id: 'speaker_28',
-    displayName: '李智樺 Ruddy',
-    avatar: '/speakers/avatar_ruddy.png',
-  },
-  {
-    id: 'speaker_29',
-    displayName: '謝舒婷 (小雨)',
-    avatar: '/speakers/avatar_lindsay.png',
-  },
-  {
-    id: 'speaker_30',
-    displayName: '林威宇',
-    avatar: '/speakers/avatar_wylin.png',
-  },
-  {
-    id: 'speaker_31',
-    displayName: '海馬',
-    avatar: '/speakers/avatar_leethi.png',
-  },
-  {
-    id: 'speaker_32',
-    displayName: 'Singz Hsieh',
-    avatar: '/speakers/avatar_singz.png',
-  },
-  {
-    id: 'speaker_33',
-    displayName: '維元',
-    avatar: '/speakers/avatar_dscareer.png',
-  },
-  {
-    id: 'speaker_34',
-    displayName: '賴怡玲 (小賴)',
-    avatar: '/speakers/avatar_lightda.png',
   },
 ]
